+++ tools/package-builder/src/core/renderer.ts
@@ -2,9 +2,10 @@ import path from 'node:path';
 import fs from 'node:fs';
 import type winston from 'winston';
 import { z } from 'zod';
-import type { Entrypoint, EntrypointType, PackageConfig, PackageEntrypoint } from './package-info';
+import type { Entrypoint, EntrypointType, PackageEntrypoint } from './package-info';
 import * as artifacts from './schemas/artifacts';
 import * as util from './util';
+import { dockerEntrypointNameToOrigin, dockerTagFromPackage } from './docker';
 
 const externalPackageLocationSchema = z.object({
   registry: z.string().describe('name of the registry to use for package download'),
@@ -12,13 +13,19 @@ const externalPackageLocationSchema = z.object({
     .string()
     .describe('full package path in registry, e.g. \'common/jdk/21.0.2.13.1-{os}-{arch}.tgz'),
 });
-type packageSwJson = z.infer<typeof externalPackageLocationSchema>;
 
 const assetSchema = z.object({
   ...externalPackageLocationSchema.shape,
   url: z.string().describe('asset download URL'),
 });
 type assetInfo = z.infer<typeof assetSchema>;
+
+const dockerSchema = z.object({
+  tag: z.string().describe('name of the image to be built instead of custom one'),
+  entrypoint: z.array(z.string()).describe('entrypoint command to be run in the container'),
+  cmd: z.array(z.string()).describe('command to be run in the container'),
+});
+type dockerInfo = z.infer<typeof dockerSchema>;
 
 const runEnvironmentSchema = z.object({
   type: z.enum(artifacts.runEnvironmentTypes),
@@ -57,9 +64,9 @@ const runDependencyCondaSchema = runEnvironmentSchema.extend({
     .string()
     .describe('name used to import this package as software dependency of tengo script'),
 });
-// type runDependencyConda = z.infer<typeof runDependencyCondaSchema>;
+type runDependencyConda = z.infer<typeof runDependencyCondaSchema>;
 
-type runDepInfo = runDependencyJava | runDependencyPython | runDependencyR;
+type runDepInfo = runDependencyJava | runDependencyPython | runDependencyR | runDependencyConda;
 // runDependencyConda
 
 const anyPackageSettingsSchema = z.object({
@@ -166,6 +173,7 @@ const entrypointSchema = z
     binary: binarySchema.optional(),
     runEnv: runEnvironmentSchema.optional(),
     local: localSchema.optional(),
+    docker: dockerSchema.optional(),
   })
   .refine(
     (data) =>
@@ -270,24 +278,16 @@ export class Renderer {
     mode: util.BuildMode,
     entrypoints: Map<string, Entrypoint>,
     options?: {
-      sources?: readonly util.SoftwareSource[];
       fullDirHash?: boolean;
     },
   ): Map<string, entrypointSwJson> {
     const result = new Map<string, entrypointSwJson>();
     let hasReferences = false;
 
-    const sources = options?.sources ?? util.AllSoftwareSources;
     const fullDirHash = options?.fullDirHash ?? false;
-
-    if (sources.length === 0) {
-      this.logger.error('list of software sources to be put into software entrypoint is empty');
-      throw new Error('nothing to render: empty list of software sources');
-    }
 
     this.logger.info(`Rendering entrypoint descriptors...`);
     this.logger.debug('  entrypoints: ' + JSON.stringify(entrypoints));
-    this.logger.debug('  sources: ' + JSON.stringify(sources));
 
     for (const [epName, ep] of entrypoints.entries()) {
       if (ep.type === 'reference') {
@@ -298,50 +298,69 @@ export class Renderer {
 
       this.logger.debug(`Rendering entrypoint descriptor '${epName}'...`);
 
-      const info: entrypointSwJson = {
-        id: {
-          package: this.npmPackageName,
-          name: epName,
-        },
-      };
+      // In docker case we should merge docker info and other info
+      //
+      const originEpName = dockerEntrypointNameToOrigin(epName);
+      const info = result.has(originEpName)
+        ? result.get(originEpName)
+        : {
+            id: {
+              package: this.npmPackageName,
+              name: originEpName,
+            },
+          };
+      if (!info) {
+        throw new Error(`Entrypoint ${epName} not found in result`);
+      }
 
       if (mode !== 'release') {
         info.isDev = true;
       }
 
-      let hasDescriptors = false;
-      for (const source of sources) {
-        switch (source) {
-          case 'archive':{
-            const pkg = ep.package;
-            if (!pkg.isBuildable) {
-              this.logger.warn('  skipping \'archive\' rendering (not a buildable package)');
-            }
-
-            hasDescriptors = true;
-            if (mode === 'dev-local') {
-              this.logger.debug('  rendering \'local\' source...');
-              info.local = this.renderLocalInfo(mode, epName, ep, fullDirHash);
-            } else {
-              this.logger.debug('  rendering \'archive\' source...');
-              if (pkg.type === 'environment') {
-                info.runEnv = this.renderRunEnvInfo(mode, epName, ep);
-              } else if (pkg.type === 'asset') {
-                info.asset = this.renderAssetInfo(mode, epName, ep);
-              } else {
-                info.binary = this.renderBinaryInfo(mode, epName, ep);
-              }
-            }
+      const pkg = ep.package;
+      if (mode === 'dev-local') {
+        switch (pkg.type) {
+          case 'docker':
+            info.docker = this.renderDockerInfo(epName, ep);
             break;
-          }
           default:
-            util.assertNever(source);
+            this.logger.debug('  rendering \'local\' source...');
+            info.local = this.renderLocalInfo(epName, ep, fullDirHash);
         }
+
+        result.set(originEpName, info);
+        continue;
       }
 
-      if (hasDescriptors) {
-        result.set(epName, info);
+      const type = pkg.type;
+      switch (type) {
+        case 'R':
+          info.binary = this.renderBinaryInfo(mode, epName, ep);
+          break;
+        case 'environment':
+          info.runEnv = this.renderRunEnvInfo(mode, epName, ep);
+          break;
+        case 'asset':
+          info.asset = this.renderAssetInfo(mode, epName, ep);
+          break;
+        case 'binary':
+          info.binary = this.renderBinaryInfo(mode, epName, ep);
+          break;
+        case 'docker':
+          info.docker = this.renderDockerInfo(epName, ep);
+          break;
+        case 'java':
+          info.binary = this.renderBinaryInfo(mode, epName, ep);
+          break;
+        case 'python':
+          info.binary = this.renderBinaryInfo(mode, epName, ep);
+          break;
+        default:
+          util.assertNever(type);
+          throw new Error(`renderer logic error: renderSoftwareEntrypoints does not cover all package types`);
       }
+
+      result.set(originEpName, info);
     }
 
     if (result.size === 0 && !hasReferences) {
@@ -350,13 +369,6 @@ export class Renderer {
     }
 
     return result;
-  }
-
-  public renderPackageDescriptor(mode: util.BuildMode, pkg: PackageConfig): packageSwJson {
-    return {
-      registry: pkg.registry.name,
-      package: pkg.namePattern,
-    };
   }
 
   public writeEntrypointDescriptor(info: entrypointSwJson, dstFile?: string) {
@@ -393,22 +405,10 @@ export class Renderer {
   }
 
   private renderLocalInfo(
-    mode: util.BuildMode,
     epName: string,
     ep: PackageEntrypoint,
     fullDirHash: boolean,
   ): localInfo {
-    switch (mode) {
-      case 'release':
-        throw new Error(`'*.sw.json' generator logic error`);
-
-      case 'dev-local':
-        break;
-
-      default:
-        util.assertNever(mode);
-    }
-
     const pkg = ep.package;
     const rootDir = pkg.contentRoot(util.currentPlatform());
     const hash = fullDirHash ? util.hashDirSync(rootDir) : util.hashDirMetaSync(rootDir);
@@ -430,12 +430,12 @@ export class Renderer {
         switch (pkgType) {
           case 'environment':{
             throw new Error(
-              `entripoint is incompatible with artifact type: ep=${epName} (software), artifact='${pkgType}'`,
+              `entrypoint is incompatible with artifact type: ep=${epName} (software), artifact='${pkgType}'`,
             );
           }
           case 'asset':{
             throw new Error(
-              `entripoint is incompatible with artifact type: ep=${epName} (software), artifact='${pkgType}'`,
+              `entrypoint is incompatible with artifact type: ep=${epName} (software), artifact='${pkgType}'`,
             );
           }
           case 'binary':{
@@ -483,6 +483,11 @@ export class Renderer {
               dependencies: {},
               runEnv: this.resolveRunEnvironment(pkg.environment, pkg.type),
             };
+          }
+          case 'docker': {
+            throw new Error(
+              `entrypoint is incompatible for local build with artifact type: ep=${epName} (software), artifact='${pkgType}'`,
+            );
           }
           // case "conda":
           //     if (runEnv!.type !== pkgType) {
@@ -554,6 +559,11 @@ export class Renderer {
           case 'environment':{
             throw new Error(
               `entripoint is incompatible with artifact type: ep=${epName} (software), artifact='${pkgType}'`,
+            );
+          }
+          case 'docker': {
+            throw new Error(
+              `entrypoint is incompatible with artifact type: ep=${epName} (software), artifact='${pkgType}'`,
             );
           }
           case 'binary':{
@@ -697,6 +707,30 @@ export class Renderer {
       registry: pkg.registry.name,
       package: pkg.namePattern,
       url: util.urlJoin(pkg.registry.downloadURL, pkg.namePattern),
+    };
+  }
+
+  private renderDockerInfo(entrypointName: string, ep: PackageEntrypoint): dockerInfo {
+    const pkg = ep.package;
+    if (pkg.type !== 'docker') {
+      throw new Error(`could not render docker entrypoint '${entrypointName}': not 'docker' artifact`);
+    }
+
+    if (ep.type !== 'software') {
+      throw new Error(`could not render docker entrypoint '${entrypointName}': not 'software' artifact`);
+    }
+
+    const blockName = this.npmPackageName.split('/').pop() ?? '';
+    if (blockName === '') {
+      throw new Error(`could not render docker entrypoint '${entrypointName}': could not determine block name from npm package name '${this.npmPackageName}'`);
+    }
+
+    const tag = dockerTagFromPackage(this.npmPackageRoot, pkg);
+
+    return {
+      tag: tag,
+      entrypoint: pkg.entrypoint ?? [],
+      cmd: ep.cmd,
     };
   }