+++ tools/pl-bootstrap/src/commands/svc/create/docker/s3.ts
@@ -5,6 +5,7 @@ import Core from '../../../../core';
 import * as cmdOpts from '../../../../cmd-opts';
 import * as util from '../../../../util';
 import state from '../../../../state';
+import { ArgParser } from '../arg-parser';
 
 export default class S3 extends Command {
   static override description = 'Run Platforma Backend service as docker container on current host with MinIO as local S3 storage';
@@ -28,18 +29,36 @@ export default class S3 extends Command {
     ...cmdOpts.MinioPresignHostFlag,
   };
 
-  static args = {
+  static override args = {
     name: Args.string({ required: true }),
   };
 
   public async run(): Promise<void> {
-    const { flags, args } = await this.parse(S3);
+    // Use custom parser instead of OCLIF parse
+    const parser = new ArgParser(S3.flags);
+    const parsed = parser.parse(this.argv);
 
-    const logger = util.createLogger(flags['log-level']);
+    // Validate required flags
+    const errors = parser.validateRequired(parsed.knownFlags);
+    if (errors.length > 0) {
+      throw new Error(`Validation errors:\n${errors.join('\n')}`);
+    }
+
+    const instanceName = parsed.instanceName;
+    const flags = parsed.knownFlags;
+
+    const backendCommands = parsed.unknownFlags;
+    if (flags['log-level']){
+      backendCommands.push(`--log-level=${flags['log-level']}`);
+    }
+
+    // Set default values for S3-specific flags
+    if (!flags['s3-port']) flags['s3-port'] = 9000;
+    if (!flags['s3-console-port']) flags['s3-console-port'] = 9001;
+
+    const logger = util.createLogger(flags['log-level'] || 'info');
     const core = new Core(logger);
     core.mergeLicenseEnvs(flags);
-
-    const instanceName = args.name;
 
     const authEnabled = flags['auth-enabled'];
     const authOptions: types.authOptions | undefined = authEnabled
@@ -84,11 +103,18 @@ export default class S3 extends Command {
       s3ConsolePort: flags['s3-console-port'],
 
       presignHost: presignHost,
+
+      backendCommands: backendCommands,
     });
 
     logger.info(`Instance '${instanceName}' was created. To start it run 'up' command`);
     if (flags['minio-presign-host']) {
       logger.info('  NOTE: make sure you have \'minio\' host in your hosts file as 127.0.0.1 address');
+    }
+
+    // Log unknown flags that will be passed to backend
+    if (backendCommands.length > 0) {
+      logger.info(`Unknown flags will be passed to backend: ${backendCommands.join(' ')}`);
     }
   }
 }