+++ tools/pl-bootstrap/src/index.ts
@@ -12,10 +12,10 @@ import Cmd8 from './commands/svc/list';
 import Cmd9 from './commands/svc/up';
 import Cmd10 from './commands/start/docker/s3';
 import Cmd11 from './commands/start/local/s3';
-import Cmd12 from './commands/svc/create/docker';
-import Cmd13 from './commands/svc/create/local';
-import Cmd14 from './commands/svc/create/docker/s3';
-import Cmd15 from './commands/svc/create/local/s3';
+import Cmd13 from './commands/svc/create/docker';
+import Cmd14 from './commands/svc/create/local';
+import Cmd15 from './commands/svc/create/docker/s3';
+import Cmd16 from './commands/svc/create/local/s3';
 
 // prettier-ignore
 export const COMMANDS = {
@@ -31,8 +31,8 @@ export const COMMANDS = {
   'svc:up': Cmd9,
   'start:docker:s3': Cmd10,
   'start:local:s3': Cmd11,
-  'svc:create:docker': Cmd12,
-  'svc:create:local': Cmd13,
-  'svc:create:docker:s3': Cmd14,
-  'svc:create:local:s3': Cmd15,
+  'svc:create:docker': Cmd13,
+  'svc:create:local': Cmd14,
+  'svc:create:docker:s3': Cmd15,
+  'svc:create:local:s3': Cmd16,
 };